backend/requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# AI Services
openai==1.59.9  # Whisper (transcription) + GPT-4 (emergency detection)

# Voice Services
elevenlabs==2.36.0
pydub==0.25.1

# Emergency & Communications
twilio==9.4.1

# Utilities
python-dotenv==1.0.1
pydantic==2.10.6  # Newer version with pre-built wheels for Python 3.13
pydantic-settings==2.7.1

# Development
pytest==8.3.5
httpx==0.28.1
